requirements.txt
Django~=5.2.3
djangorestframework~=3.16.0
requests~=2.32.4
beautifulsoup4~=4.13.4
soupsieve~=2.7
packaging~=25.0
typing_extensions~=4.14.0
bleach~=6.2.0
pip~=25.0.1
attrs~=25.3.0
pyOpenSSL~=25.1.0
cryptography~=45.0.4
amqp~=5.3.1
vine~=5.1.0
kombu~=5.5.4
billiard~=4.2.1
urllib3~=2.4.0
celery~=5.5.3
click~=8.2.1
python-dateutil~=2.9.0.post0
click-repl~=0.3.0
Pygments~=2.19.1
click-plugins~=1.1.1
psutil~=7.0.0
bson~=0.5.10
asgiref~=3.8.1
sqlparse~=0.5.3
psycopg2~=2.9.10
colorama~=0.4.6
PyYAML~=6.0.2
Jinja2~=3.1.5
numpy~=2.2.0
certifi~=2025.6.15
wcwidth~=0.2.13
cffi~=1.17.1
six~=1.17.0
idna~=3.10
charset-normalizer~=3.4.2
chardet~=3.0.4
prompt_toolkit~=3.0.48
setuptools~=80.9.0
pytz~=2024.2
dotenv~=0.9.9
python-dotenv~=1.1.0